backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
google-generativeai
jinja2